requirements.txt
python-frontmatter
wordpress-api
requests-oauthlib
